pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lab-more-project-chat"
version = "0.1.0"
description = "A local project chat assistant with safe file tools and slash commands."
readme = "README.md"
license = "MIT"
authors = [{name = "Mia U"}]
requires-python = ">=3.10"
dependencies = ["requests", "GitPython"]

[tool.setuptools]
py-modules = ["chat"]

[tool.setuptools.packages.find]
include = ["tools"]

[project.scripts]
chat = "chat:main"
docchat = "chat:main"
